encapsulate home component
